module8-task1: refactoring namespaces, slices, actions
