+++ src/components/catalog/catalog.tsx
@@ -3,7 +3,7 @@ import GenresList from '../genres-list/genres-list.tsx';
 import React, { useCallback, useState } from 'react';
 import { useAppSelector } from '../../hooks/store.ts';
 import { DEFAULT_FILM_LIST_LENGTH } from '../../consts/film-list.ts';
-import { getFilmsByGenreLength } from '../../store/film-process/film-process.selector.ts';
+import { getFilmsByGenreLength } from '../../store/films-process/films-process.selector.ts';
 
 export default function Catalog(): React.JSX.Element {
   const stateGenreFilmsLength = useAppSelector(getFilmsByGenreLength);
