+++ src/pages/main/main.tsx
@@ -4,7 +4,7 @@ import FilmCard from '../../components/film-card/film-card';
 import Catalog from '../../components/catalog/catalog';
 import { useAppDispatch, useAppSelector } from '../../hooks/store.ts';
 import { fetchFilmPromo } from '../../store/api-actions.ts';
-import { getPromoFilm } from '../../store/film-process/film-process.selector.ts';
+import { getPromoFilm } from '../../store/films-process/films-process.selector.ts';
 
 export default function MainPage(): React.JSX.Element | null {
   const dispatch = useAppDispatch();
