+++ src/types/auth.ts
@@ -7,4 +7,6 @@ export type UserData = {
   id: number;
   email: string;
   token: string;
+  name: string;
+  avatarUrl: string;
 }
